fix: 비동기 메서드 다른 클래스로 분리
