+++ PJA/src/main/java/com/project/PJA/PjaApplication.java
@@ -3,8 +3,11 @@ package com.project.PJA;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.scheduling.annotation.EnableAsync;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
-//@EnableScheduling
+@EnableAsync
+@EnableScheduling
 @SpringBootApplication
 @EntityScan(basePackages = "com.project.PJA")
 public class PjaApplication {
